remove unused legacy code
