+++ src/mercs/graph/__init__.py
@@ -1,15 +1,3 @@
-from .network import (
-    model_to_graph,
-    add_merge_nodes,
-    add_imputation_nodes,
-    compose,
-    get_ids,
-    get_nodes,
-    compose_all,
-    get_targ,
-    get_desc
-)
-
 from .graphviz import fix_layout, to_dot
 
 from .q_diagram import build_diagram
